refactor: rename pr-status-line to pr-status-model
