+++ fab/changes/260719-8hc6-pr-status-model-rename/.status.yaml
@@ -1,0 +1,55 @@
+id: 8hc6
+name: 260719-8hc6-pr-status-model-rename
+created: 2026-07-19T20:20:51Z
+created_by: sahil-noon
+change_type: refactor
+issues: []
+progress:
+    intake: done
+    apply: done
+    review: done
+    hydrate: done
+    ship: active
+    review-pr: pending
+plan:
+    generated: true
+    task_count: 11
+    acceptance_count: 8
+    acceptance_completed: 8
+confidence:
+    certain: 2
+    confident: 2
+    tentative: 0
+    unresolved: 0
+    score: 4.9
+    fuzzy: true
+    dimensions:
+        signal: 70.0
+        reversibility: 88.8
+        competence: 88.8
+        disambiguation: 81.3
+stage_metrics:
+    intake: {started_at: "2026-07-19T20:20:51Z", driver: fab-new, iterations: 1, completed_at: "2026-07-19T20:21:39Z"}
+    apply: {started_at: "2026-07-19T20:21:39Z", driver: fab-fff, iterations: 1, completed_at: "2026-07-19T20:25:04Z"}
+    review: {started_at: "2026-07-19T20:25:04Z", driver: fab-fff, iterations: 1, completed_at: "2026-07-19T20:28:43Z"}
+    hydrate: {started_at: "2026-07-19T20:28:43Z", driver: fab-fff, iterations: 1, completed_at: "2026-07-19T20:31:43Z"}
+    ship: {started_at: "2026-07-19T20:31:43Z", driver: fab-fff, iterations: 1}
+prs: []
+change_type_source: explicit
+true_impact:
+    added: 0
+    deleted: 0
+    net: 0
+    excluding:
+        added: 0
+        deleted: 0
+        net: 0
+    tests:
+        added: 0
+        deleted: 0
+        net: 0
+    computed_at: "2026-07-19T20:31:43Z"
+    computed_at_stage: hydrate
+summary: Repointed run-kit/ui-patterns status-dot-model module path from pr-status-line.tsx to pr-status-model.ts (rename)
+# true_impact: lazily created on first stage-finish that computes it (no placeholder here).
+last_updated: 2026-07-19T20:31:43Z
